GTD with org mode; revert to keeping elpa, ask to update; slime issues
Main change is that I'm now using org mode for my GTD setup. This configuration
is changing rapidly though.

Elpa (GNU repo) is back to being used, in addition to Melpa, and Emacs asks if
you want to check for new packaages before doing so.

After restarting Emacs with the current version of Slime, I noticed issues with
it not finding the hyperspec in the lib directory, so have had to temporarily
comment out several slime, common lisp-related lines.